storage: update BlobPrefetchRequest fields type from u32 to u64
Blobfs relies on the BlobPrefetchRequest structure that field of "offset"
and "len" are both u32, the maximum support is 4G. But the maximum blob file
will exceed 4G, so change the define of "offset" and "len" to u64.

+++ storage/src/backend/oss.rs
@@ -202,7 +202,7 @@ impl BlobReader for OssReader {
             .map(|size| size as usize)?)
     }
 
-    fn prefetch_blob_data_range(&self, _ra_offset: u32, _ra_size: u32) -> BackendResult<()> {
+    fn prefetch_blob_data_range(&self, _ra_offset: u64, _ra_size: u64) -> BackendResult<()> {
         Err(BackendError::Unsupported(
             "Oss backend does not support prefetch as per on-disk blob entries".to_string(),
         ))
